flask-pcolor/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
blis==0.7.7
cached-property==1.5.2
cachetools==5.0.0
catalogue==2.0.7
certifi==2019.11.28
charset-normalizer==2.0.12
click==8.0.4
cmake==3.22.4
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
efficientnet-pytorch==0.7.1
fastai==2.5.6
fastcore==1.4.2
fastdownload==0.0.5
fastprogress==1.0.2
Flask==2.1.1
flatbuffers==2.0
fonttools==4.33.2
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
langcodes==3.3.0
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
murmurhash==1.0.6
numpy @ file:///D:/bld/numpy_1649806536290/work
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathy==0.6.1
Pillow==9.1.0
preshed==3.0.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.15
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typer==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.14.0
zipp==3.8.0
